+++ packages/diatrema/src/index.ts
@@ -1,6 +1,0 @@
-export * from './diatrema.js';
-export * from './event-emitter.js';
-
-import { Diatrema } from './diatrema.js';
-
-export default Diatrema;